Renamed money package
Because it has many providers.
