+++ commercial/app/model/commercial/money/BestBuys.scala
@@ -1,4 +1,4 @@
-package model.commercial.moneysupermarket
+package model.commercial.money
 
 import model.commercial.Segment
