+++ commercial/app/views/moneysupermarket/bestBuys.scala.html
@@ -1,4 +1,4 @@
-@(bestBuys: model.commercial.moneysupermarket.BestBuys)(implicit request: RequestHeader)
+@(bestBuys: model.commercial.money.BestBuys)(implicit request: RequestHeader)
 
 <link rel="stylesheet" href="@Static("stylesheets/commercial.css")" />
