+++ commercial/test/model/commercial/money/CurrentAccountsApiTest.scala
@@ -1,4 +1,4 @@
-package model.commercial.moneysupermarket
+package model.commercial.money
 
 import org.scalatest.{Matchers, FlatSpec}
